fix: refactor `mutate-when-known-kind` into its own folder and structure / remove deprectated functions
