--- clients/web/src/lib/span/is-event-span.ts
@@ -1,17 +1,0 @@
-import { Span } from "../connection/monitor";
-
-const emitSpanNames = [
-  "app::emit",
-  "app::emit::all",
-  "app::emit::filter",
-  "app::emit::to",
-  "app::emit::rust",
-  "window::emit",
-  "window::trigger",
-  "window::emit::to",
-  "window::emit::all",
-];
-
-export function isEventSpan(span: Span) {
-  return emitSpanNames.includes(span.metadata?.name ?? "");
-}
