+++ clients/web/src/lib/span/update/kinds/event/event-spans.ts
@@ -1,6 +1,4 @@
-import type { Span, EventKind } from "~/lib/connection/monitor";
-
-const eventTypes = [
+export const EventKinds = [
   {
     type: "global event",
     names: new Set([
@@ -20,13 +18,3 @@ const eventTypes = [
     names: new Set(["window::emit", "window::emit::to", "window::emit::all"]),
   },
 ] as const;
-
-export function detectEventKind(span: Span): EventKind | undefined {
-  if (!span.metadata?.name) return;
-
-  for (const eventType of eventTypes) {
-    if (eventType.names.has(span.metadata.name)) {
-      return eventType.type;
-    }
-  }
-}
