+++ clients/web/src/lib/span/update/kinds/event/mutate-when-event-trace.ts
@@ -1,0 +1,11 @@
+import type { Span } from "~/lib/connection/monitor";
+import { detectEventTrace } from "./detect-event-trace";
+
+export function mutateWhenEventTrace(root: Span): boolean {
+  const eventData = detectEventTrace(root);
+  if (!eventData) return false;
+  root.kind = "event";
+  root.displayName = `${eventData.kind}: ${eventData.event}`;
+  root.eventData = eventData;
+  return true;
+}
